requirements.txt
Django==4.0.3
django-ckeditor==6.2.0
django-js-asset==2.0.0
gunicorn==20.1.0
psycopg2==2.9.3
whitenoise==6.0.0
